Course
Upadete course funciotns delete by ID and delete all for doesnt throw error 500

+++ src/main/java/com/project/professor/allocation/service/CourseService.java
@@ -9,7 +9,9 @@ import org.springframework.stereotype.Service;
 import com.project.professor.allocation.entity.Course;
 import com.project.professor.allocation.entity.Department;
 import com.project.professor.allocation.entity.Professor;
+import com.project.professor.allocation.repository.AllocationRepository;
 import com.project.professor.allocation.repository.CourseRepository;
+import com.project.professor.allocation.service.exception.AllocatedExistsException;
 import com.project.professor.allocation.service.exception.EntityNotFindException;
 
 import net.bytebuddy.implementation.bytecode.Throw;
@@ -18,10 +20,12 @@ import net.bytebuddy.implementation.bytecode.Throw;
 public class CourseService {
 
 	private final CourseRepository courseRepository;
+	private final AllocationRepository allocationRepository;
 
-	public CourseService(CourseRepository courseRepository) {
+	public CourseService(CourseRepository courseRepository, AllocationRepository allocationRepository) {
 		super();
 		this.courseRepository = courseRepository;
+		this.allocationRepository = allocationRepository;
 	}
 
 	public Course findByCourseId(Long courseId) throws EntityNotFindException {
@@ -61,15 +65,28 @@ public class CourseService {
 		}
 	}
 
-	public void deleteById(Long id) throws EntityNotFindException {
+	public void deleteById(Long id) throws EntityNotFindException, AllocatedExistsException {
 		if (id != null && courseRepository.existsById(id)) {
-			courseRepository.deleteById(id);
+			if(allocationRepository.findByCourseId(id) == null) {
+				courseRepository.deleteById(id);	
+			}else {
+				throw new AllocatedExistsException("This course have allocation");
+			}
+			
 		} else {
 			throw new EntityNotFindException("Course doesn't find");
 		}
 	}
 
-	public void deleteAll() {
+	public void deleteAll() throws AllocatedExistsException {
+		
+		List<Course> courses = findAll();
+		
+		for (Course course : courses) {
+			if(allocationRepository.findByCourseId(course.getId()) != null) {
+				throw new AllocatedExistsException("Course have allocation");	
+			}
+		}
 		courseRepository.deleteAllInBatch();
 	}
 }
